@@ -1,10 +1,14 @@
+import { useParams } from "react-router-dom";
+
 import Layout from "./Layout";
 import Hotel from "../components/Hotel";
 
-const HotelView = (props) => {
+const HotelView = () => {
+    const { id } = useParams();
+
     return (
-        <Layout history={props.history}>
-            <Hotel id={props.match.params.id} history={props.history} />
+        <Layout>
+            <Hotel id={id} />
         </Layout>
     );
 };
