@@ -1,13 +1,13 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
-import mdPlugin, { Mode } from "vite-plugin-markdown";
+import { plugin as mdPlugin, Mode } from "vite-plugin-markdown";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react(), mdPlugin({ mode: Mode.HTML })],
-  test: {
-    globals: true,
-    environment: "jsdom",
-    restoreMocks: true,
-  },
+    plugins: [react(), mdPlugin({ mode: Mode.HTML })],
+    test: {
+        globals: true,
+        environment: "jsdom",
+        restoreMocks: true,
+    },
 });
